Move RestrictedVocabularyFactory to opengever.base.restricted_vocab
